+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - The GUI can add components to entities
 - Made it work for web again
 - Smooth damp
+- More options in the menu
 #### Changed
 - Improved palette handling and added new ones
 - Randomized palette on start
@@ -18,10 +19,11 @@
 - Now GL_LUMINANCE is used for WebGL instead of making RGBA textures
 - Noise component
 - Made camera a bit more static, but solved the brrrr issue.
-- More thinking into how to apply the smooth damp only when the player is still is needed.
+- Extra and smooth  lookahead
 #### Fixed
 - When going down on platform you went through solid blocks
 - Fixed some issues with perlin noise, normalized the values
+- FIXED THE CAMERA, FINALLY :DDD
 
 ---